+++ CHANGELOG.md
@@ -25,3 +25,7 @@
 - Found: step_1_data_preparation/step1.2/._validate_input_data.py
 - Found: step_1_data_preparation/step1.2/.fuse_hidden0002b87900000001
 - Found: step_1_data_preparation/step1.2/LMOD_SETUP_INSTRUCTIONS.html
+
+## [2025-11-13 14:36:38] Auto Commit
+
+- Found: step_1_data_preparation/step1.3/